+++ src/main/java/nextstep/subway/entity/Line.java
@@ -15,42 +15,17 @@ public class Line {
     @Column(length = 20, nullable = false)
     private String color;
 
-    @Column(nullable = false)
-    private Integer distance;
-
     @Embedded
     private Sections sections = new Sections();
-
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(
-        name = "up_staion_id",
-        foreignKey = @ForeignKey(ConstraintMode.NO_CONSTRAINT),
-        nullable = false
-    )
-    private Station upStation;
-
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(
-        name = "down_staion_id",
-        foreignKey = @ForeignKey(ConstraintMode.NO_CONSTRAINT),
-        nullable = false
-    )
-    private Station downStation;
 
     protected Line() {}
 
     public Line(
         String name,
-        String color,
-        Integer distance,
-        Station upStation,
-        Station downStation
+        String color
     ) {
         this.name = name;
         this.color = color;
-        this.distance = distance;
-        this.upStation = upStation;
-        this.downStation = downStation;
     }
 
     public void updateLine(String name, String color) {
@@ -78,20 +53,9 @@ public class Line {
         return color;
     }
 
-    public Integer getDistance() {
-        return distance;
-    }
 
     public Sections getSections() {
         return sections;
-    }
-
-    public Station getUpStation() {
-        return upStation;
-    }
-
-    public Station getDownStation() {
-        return downStation;
     }
 
     public List<Section> getSectionList() {
