+++ src/test/java/nextstep/subway/unit/LineServiceMockTest.java
@@ -38,7 +38,7 @@ public class LineServiceMockTest {
     void setUp() {
         강남역 = new Station("강남역");
         역삼역 = new Station("역삼역");
-        이호선 = new Line("2호선", "green", 10, 강남역, 역삼역);
+        이호선 = new Line("2호선", "green");
     }
 
     @Test
